removed device_helpers.go and moved helpers to their modals
